Fixed route path for selection and refactored MenuMain.jsx into its own folder
