+++ src/components/Routes.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Route, Routes } from 'react-router-dom'
-import MenuMain from './MenuMain'
+import MainMenu from './mainmenu'
 
 import Home from './home/index'
 
@@ -12,8 +12,7 @@ function RoutesList(props) {
   return (
     <Routes>
         <Route path={appConstants.routes.home} element={<Home/>}/>
-        {/* TODO: replace the path with a constant from appConstants */}
-        <Route path="/select" element={<MenuMain />} />
+        <Route path={appConstants.routes.gameChoice} element={<MainMenu/>} />
         <Route path={appConstants.routes.notFound} element={<h1 className='text-red-600 font-bold'>404 Not Found</h1>}/>
     </Routes>
   )
